Refactor + test with synapso-core
