+++ .gitignore
@@ -172,5 +172,5 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
-
+.cursor
 .DS_Store
