+++ src/synapso_cli/commands/cortex/__init__.py
@@ -1,3 +1,3 @@
-from .base import cortex_app
+from .commands import cortex_app
 
 __all__ = ["cortex_app"]
