+++ src/synapso_cli/commands/cortex/commands.py
@@ -1,8 +1,5 @@
 import typer
-
-from .create import cmd_create_cortex
-from .index import cmd_index_cortex
-from .init import cmd_initialize_cortex
+from synapso_core.cortex_manager import CortexManager
 
 cortex_app = typer.Typer()
 
@@ -24,20 +21,6 @@ def create(
     )
 
 
-@cortex_app.command(name="init")
-def initialize(
-    cortex_id: str = typer.Option(
-        ..., help="The ID of the cortex to initialize", metavar="CORTEX_ID"
-    ),
-    index_now: bool = typer.Option(
-        False, help="Whether to index the cortex now", metavar="INDEX_NOW"
-    ),
-):
-    cmd_initialize_cortex(cortex_id, index_now)
-    action = "initialized and indexed" if index_now else "initialized"
-    typer.echo(f"Cortex {cortex_id} {action} successfully")
-
-
 @cortex_app.command(name="index")
 def index(
     cortex_id: str = typer.Option(
@@ -46,3 +29,25 @@ def index(
 ):
     cmd_index_cortex(cortex_id)
     typer.echo(f"Cortex {cortex_id} indexed successfully")
+
+
+def cmd_create_cortex(folder_location: str, cortex_name: str):
+    cortex_manager = CortexManager()
+    return cortex_manager.create_cortex(cortex_name, folder_location)
+
+
+def cmd_initialize_cortex(cortex_id: str, index_now: bool = False):
+    cortex_manager = CortexManager()
+    return cortex_manager.initialize_cortex(cortex_id, index_now)
+
+
+def cmd_index_cortex(cortex_id: str):
+    cortex_id = cortex_id.strip().lower()
+    if not cortex_id:
+        raise typer.BadParameter("Cortex ID is required")
+    try:
+        cortex_manager = CortexManager()
+        return cortex_manager.index_cortex(cortex_id)
+    except Exception as e:
+        typer.echo(f"Error indexing cortex: {e}", err=True)
+        raise typer.Exit(1) from e
