--- src/synapso_cli/commands/cortex/init.py
@@ -1,6 +1,0 @@
-from synapso_core.cortex_manager import CortexManager
-
-
-def cmd_initialize_cortex(cortex_id: str, index_now: bool = False):
-    cortex_manager = CortexManager()
-    return cortex_manager.initialize_cortex(cortex_id, index_now)
